// src/app/common/fhir/fhir.resource.type.ts
export enum ResourceType {
    Attachment = 'Attachment',
    Bundle = 'Bundle',
    Patient = 'Patient',
    Practitioner = 'Practitioner',
    Library = 'Library',
    Parameters = 'Parameters',
    MedicationRequest = 'MedicationRequest',
    MedicationKnowledge = 'MedicationKnowledge',
    Medication = 'Medication',
}